Update React Native types and app version; upgrading Babel & core-js dependencies
Updated the '@types/react-native' version from '~0.66.13' to '~0.67.6' in package.json and package-lock.json files to be compatible with the latest changes in the React Native framework.
Also, the version in app.json was updated from '1.0.0' to '1.0.1' to reflect the changes in the application.
Moreover, several '@babel/*' dependencies and 'core-js-compat' were updated in package-lock.json, including the resolution URLs and integrity hashes, to utilize the latest versions and corresponding enhancements.